fix site layout
