+++ components/snippet-chooser/index.js
@@ -1,10 +1,3 @@
-import { connect } from 'react-redux'
 import SnippetChooser from './view'
-import { updateSnippet } from '../../store/actions'
 
-export const mapDispatchToProps = dispatch => ({
-  onButtonClick: snippet => {
-    dispatch(updateSnippet(snippet))
-  },
-})
-export default connect(null, mapDispatchToProps)(SnippetChooser)
+export default SnippetChooser
